新增类QQAIPicClass以封装参数相同的3个功能
